Refactor admin class analytics charts component
