+++ frontend/src/pages/dashboard/admin/online-classes/[id]/analytics.js
@@ -9,7 +9,10 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import nextI18NextConfig from "../../../../../../next-i18next.config.js";
 import { fetchAdminClassAnalytics } from "@/services/admin/classService";
 
-const COLORS = ["#60a5fa", "#34d399", "#fbbf24", "#f87171"];
+const AnalyticsCharts = dynamic(
+  () => import("@/components/admin/online-classes/AnalyticsCharts"),
+  { ssr: false }
+);
 
 // ─────────────────────
 // Fallback analytics when API fails
@@ -34,37 +37,6 @@ function AnalyticsDashboard() {
   const { id } = router.query;
   const { t, i18n } = useTranslation('dashboard');
   const [stats, setStats] = useState(null);
-  const [chartsLib, setChartsLib] = useState(null);
-  const [resizeObserverSupported, setResizeObserverSupported] = useState(true);
-  const [chartsLoadError, setChartsLoadError] = useState(false);
-
-  useEffect(() => {
-    if (typeof window === "undefined") {
-      return;
-    }
-
-    if (!window.ResizeObserver) {
-      setResizeObserverSupported(false);
-      return;
-    }
-
-    let isMounted = true;
-    import("recharts")
-      .then((module) => {
-        if (!isMounted) return;
-        setChartsLib(module);
-      })
-      .catch((error) => {
-        console.error("Failed to load Recharts", error);
-        if (!isMounted) return;
-        setChartsLoadError(true);
-        setChartsLib(null);
-      });
-
-    return () => {
-      isMounted = false;
-    };
-  }, []);
 
   useEffect(() => {
     if (!id) return;
@@ -104,17 +76,6 @@ function AnalyticsDashboard() {
   const locations = stats.locations ?? EMPTY_STATS.locations;
   const devices = stats.devices ?? EMPTY_STATS.devices;
   const registrationTrend = stats.registrationTrend ?? EMPTY_STATS.registrationTrend;
-  const ResponsiveContainer = chartsLib?.ResponsiveContainer;
-  const PieChart = chartsLib?.PieChart;
-  const Pie = chartsLib?.Pie;
-  const Cell = chartsLib?.Cell;
-  const Legend = chartsLib?.Legend;
-  const Tooltip = chartsLib?.Tooltip;
-  const BarChart = chartsLib?.BarChart;
-  const Bar = chartsLib?.Bar;
-  const XAxis = chartsLib?.XAxis;
-  const YAxis = chartsLib?.YAxis;
-  const CartesianGrid = chartsLib?.CartesianGrid;
 
   const avgRevenue =
     totalStudents > 0
@@ -173,98 +134,12 @@ function AnalyticsDashboard() {
         </div>
       </div>
 
-      <div className="grid md:grid-cols-2 gap-6">
-        <div className="bg-white p-6 rounded-xl shadow border">
-          <h2 className="text-lg font-semibold text-gray-700 mb-4">🌍 {t('classAnalyticsPage.top_countries')}</h2>
-          {resizeObserverSupported && ResponsiveContainer && PieChart ? (
-            <ResponsiveContainer width="100%" height={300}>
-              <PieChart>
-                <Pie data={locations} dataKey="value" nameKey="name" outerRadius={100}>
-                  {(locations ?? []).map((entry, index) => (
-                    <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-                  ))}
-                </Pie>
-                <Legend />
-                <Tooltip />
-              </PieChart>
-            </ResponsiveContainer>
-          ) : (
-            <div className="flex h-[300px] items-center justify-center px-6 text-center text-sm text-muted-foreground">
-              {resizeObserverSupported
-                ? chartsLoadError
-                  ? t(
-                      'classAnalyticsPage.chartsFailedToLoad',
-                      'Charts failed to load. Please refresh to try again.'
-                    )
-                  : t('classAnalyticsPage.loadingCharts', 'Loading charts…')
-                : t(
-                    'classAnalyticsPage.chartsUnavailableResizeObserver',
-                    'Charts are unavailable because ResizeObserver is not supported in this browser.'
-                  )}
-            </div>
-          )}
-        </div>
-
-        <div className="bg-white p-6 rounded-xl shadow border">
-          <h2 className="text-lg font-semibold text-gray-700 mb-4">📱 {t('classAnalyticsPage.devices_used')}</h2>
-          {resizeObserverSupported && ResponsiveContainer && PieChart ? (
-            <ResponsiveContainer width="100%" height={300}>
-              <PieChart>
-                <Pie data={devices} dataKey="value" nameKey="name" outerRadius={100}>
-                  {(devices ?? []).map((entry, index) => (
-                    <Cell key={`cell-dev-${index}`} fill={COLORS[index % COLORS.length]} />
-                  ))}
-                </Pie>
-                <Legend />
-                <Tooltip />
-              </PieChart>
-            </ResponsiveContainer>
-          ) : (
-            <div className="flex h-[300px] items-center justify-center px-6 text-center text-sm text-muted-foreground">
-              {resizeObserverSupported
-                ? chartsLoadError
-                  ? t(
-                      'classAnalyticsPage.chartsFailedToLoad',
-                      'Charts failed to load. Please refresh to try again.'
-                    )
-                  : t('classAnalyticsPage.loadingCharts', 'Loading charts…')
-                : t(
-                    'classAnalyticsPage.chartsUnavailableResizeObserver',
-                    'Charts are unavailable because ResizeObserver is not supported in this browser.'
-                  )}
-            </div>
-          )}
-        </div>
-      </div>
-
-      <div className="bg-white p-6 rounded-xl shadow border">
-        <h2 className="text-lg font-semibold text-gray-700 mb-4">📈 {t('classAnalyticsPage.registration_trend')}</h2>
-        {resizeObserverSupported && ResponsiveContainer && BarChart ? (
-          <ResponsiveContainer width="100%" height={300}>
-            <BarChart data={registrationTrend}>
-              <CartesianGrid strokeDasharray="3 3" />
-              <XAxis dataKey="date" />
-              <YAxis />
-              <Tooltip />
-              <Bar dataKey="students" fill="#facc15" />
-            </BarChart>
-          </ResponsiveContainer>
-        ) : (
-          <div className="flex h-[300px] items-center justify-center px-6 text-center text-sm text-muted-foreground">
-            {resizeObserverSupported
-              ? chartsLoadError
-                ? t(
-                    'classAnalyticsPage.chartsFailedToLoad',
-                    'Charts failed to load. Please refresh to try again.'
-                  )
-                : t('classAnalyticsPage.loadingCharts', 'Loading charts…')
-              : t(
-                  'classAnalyticsPage.chartsUnavailableResizeObserver',
-                  'Charts are unavailable because ResizeObserver is not supported in this browser.'
-                )}
-          </div>
-        )}
-      </div>
+      <AnalyticsCharts
+        t={t}
+        locations={locations}
+        devices={devices}
+        registrationTrend={registrationTrend}
+      />
     </div>
   );
 }
